chore: clean aftermarth of FWUpdate and fix functional components
