+++ src/renderer/App.js
@@ -35,7 +35,6 @@ import Dark from "./theme/DarkTheme";
 
 import SelectKeyboard from "./views/SelectKeyboard";
 import FirmwareUpdate from "./views/FirmwareUpdate";
-import AltFirmwareUpdate from "./views/AltFirmwareUpdate";
 import LayoutEditor from "./views/LayoutEditor";
 import MacroEditor from "./views/MacroEditor";
 import SuperkeysEditor from "./views/SuperkeysEditor";
@@ -390,7 +389,7 @@ class App extends React.Component {
               inContext={contextBar}
               titleElement={() => document.querySelector("#page-title")}
             />
-            <AltFirmwareUpdate
+            <FirmwareUpdate
               path="/firmware-update"
               device={this.state.device}
               toggleFlashing={this.toggleFlashing}
