+++ src/renderer/modules/Firmware/FirmwareErrorPanel.js
@@ -199,7 +199,7 @@ width: 100%;
  * @returns {<FirmwareUpdatePanel>} FirmwareUpdatePanel component.
  */
 
-const FirmwareErrorPanel = ({ nextBlock, retryBlock }) => {
+function FirmwareErrorPanel({ nextBlock, retryBlock }) {
   const [state, send] = useMachine(FWSelection);
   const [handleError, setHandleError] = useState(false);
 
@@ -290,6 +290,6 @@ const FirmwareErrorPanel = ({ nextBlock, retryBlock }) => {
       <div style={{ display: "none" }}>{JSON.stringify(state.context)}</div>
     </Style>
   );
-};
+}
 
 export default FirmwareErrorPanel;
